app/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==1.4.46
psycopg2-binary==2.9.6
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
supabase==0.5.8
alembic==1.10.3
python-dotenv==1.0.0
bcrypt==4.0.1
pytest==7.3.1
httpx==0.24.0
requests==2.30.0